feat: split genesis config from config.go to genesis.go
