+++ internal/finance/gas.go
@@ -34,8 +34,8 @@ func NewGas(repo *repo.Repo) *Gas {
 //
 // if G_c >= maxGasPrice, G_c = maxGasPrice
 func (gas *Gas) CalNextGasPrice(parentGasPrice uint64, txs int) (uint64, error) {
-	max := gas.repo.Config.Genesis.EpochInfo.FinanceParams.MaxGasPrice
-	min := gas.repo.Config.Genesis.EpochInfo.FinanceParams.MinGasPrice
+	max := gas.repo.GenesisConfig.EpochInfo.FinanceParams.MaxGasPrice
+	min := gas.repo.GenesisConfig.EpochInfo.FinanceParams.MinGasPrice
 	if parentGasPrice < min || parentGasPrice > max {
 		gas.logger.Errorf("gas price is out of range, parent gas price is %d, min is %d, max is %d", parentGasPrice, min, max)
 		return 0, ErrGasOutOfRange
@@ -44,7 +44,7 @@ func (gas *Gas) CalNextGasPrice(parentGasPrice uint64, txs int) (uint64, error) 
 	if txs > total {
 		return 0, ErrTxsOutOfRange
 	}
-	percentage := 2 * float64(txs-total/2) / float64(total) * float64(gas.repo.Config.Genesis.EpochInfo.FinanceParams.GasChangeRateValue) / math.Pow10(int(gas.repo.Config.Genesis.EpochInfo.FinanceParams.GasChangeRateDecimals))
+	percentage := 2 * float64(txs-total/2) / float64(total) * float64(gas.repo.GenesisConfig.EpochInfo.FinanceParams.GasChangeRateValue) / math.Pow10(int(gas.repo.GenesisConfig.EpochInfo.FinanceParams.GasChangeRateDecimals))
 	currentPrice := uint64(float64(parentGasPrice) * (1 + percentage))
 	if currentPrice > max {
 		gas.logger.Warningf("gas price is touching ceiling, current price is %d, max is %d", currentPrice, max)
