+++ internal/ledger/genesis/genesis.go
@@ -14,7 +14,7 @@ import (
 	"github.com/axiomesh/axiom-ledger/pkg/repo"
 )
 
-func initializeGenesisConfig(genesis *repo.Genesis, lg ledger.StateLedger) error {
+func initializeGenesisConfig(genesis *repo.GenesisConfig, lg ledger.StateLedger) error {
 	account := lg.GetOrCreateAccount(types.NewAddressByStr(common.ZeroAddress))
 
 	genesisCfg, err := json.Marshal(genesis)
@@ -26,7 +26,7 @@ func initializeGenesisConfig(genesis *repo.Genesis, lg ledger.StateLedger) error
 }
 
 // Initialize initialize block
-func Initialize(genesis *repo.Genesis, lg *ledger.Ledger) error {
+func Initialize(genesis *repo.GenesisConfig, lg *ledger.Ledger) error {
 	dummyRootHash := common2.Hash{}
 	lg.StateLedger.PrepareBlock(types.NewHash(dummyRootHash[:]), nil, 1)
