+++ pkg/repo/repo.go
@@ -27,6 +27,7 @@ type Repo struct {
 	RepoRoot        string
 	Config          *Config
 	ConsensusConfig *ConsensusConfig
+	GenesisConfig   *GenesisConfig
 	AccountKey      *ecdsa.PrivateKey
 	AccountAddress  string
 	P2PKey          *ecdsa.PrivateKey
@@ -145,18 +146,21 @@ func DefaultWithNodeIndex(repoRoot string, nodeIndex int, epochEnable bool) (*Re
 		return nil, err
 	}
 
-	cfg := DefaultConfig(epochEnable)
+	cfg := DefaultConfig()
 	cfg.Port.P2P = int64(4001 + nodeIndex)
+
+	genesisCfg := DefaultGenesisConfig(epochEnable)
 
 	return &Repo{
 		RepoRoot:        repoRoot,
 		Config:          cfg,
 		ConsensusConfig: DefaultConsensusConfig(),
+		GenesisConfig:   genesisCfg,
 		AccountKey:      accountKey,
 		AccountAddress:  ethcrypto.PubkeyToAddress(accountKey.PublicKey).String(),
 		P2PKey:          p2pKey,
 		P2PID:           id,
-		EpochInfo:       cfg.Genesis.EpochInfo,
+		EpochInfo:       genesisCfg.EpochInfo,
 	}, nil
 }
 
@@ -190,15 +194,21 @@ func Load(repoRoot string) (*Repo, error) {
 		return nil, err
 	}
 
+	genesisCfg, err := LoadGenesisConfig(repoRoot)
+	if err != nil {
+		return nil, err
+	}
+
 	repo := &Repo{
 		RepoRoot:        repoRoot,
 		Config:          cfg,
 		ConsensusConfig: consensusCfg,
+		GenesisConfig:   genesisCfg,
 		AccountKey:      accountKey,
 		AccountAddress:  ethcrypto.PubkeyToAddress(accountKey.PublicKey).String(),
 		P2PKey:          p2pKey,
 		P2PID:           id,
-		EpochInfo:       cfg.Genesis.EpochInfo,
+		EpochInfo:       genesisCfg.EpochInfo,
 	}
 
 	return repo, nil
