+++ scripts/package/.env.sh
@@ -10,20 +10,21 @@ set -e
 
 
 # genesis
-#export AXIOM_LEDGER_GENESIS_CHAINID=1356
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_P2P_BOOTSTRAP_NODE_ADDRESSES="\
+#export AXIOM_LEDGER_CHAINID=1356
+#export AXIOM_LEDGER_EPOCH_INFO_P2P_BOOTSTRAP_NODE_ADDRESSES="\
 #/ip4/127.0.0.1/tcp/4001/p2p/16Uiu2HAmJ38LwfY6pfgDWNvk3ypjcpEMSePNTE6Ma2NCLqjbZJSF;\
 #/ip4/127.0.0.1/tcp/4002/p2p/16Uiu2HAmRypzJbdbUNYsCV2VVgv9UryYS5d7wejTJXT73mNLJ8AK;\
 #/ip4/127.0.0.1/tcp/4003/p2p/16Uiu2HAmTwEET536QC9MZmYFp1NUshjRuaq5YSH1sLjW65WasvRk;\
 #/ip4/127.0.0.1/tcp/4004/p2p/16Uiu2HAmQBFTnRr84M3xNhi3EcWmgZnnBsDgewk4sNtpA3smBsHJ;\
 #"
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_EPOCH_PERIOD=100
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_CONSENSUS_PARAMS_CHECKPOINT_PERIOD=1
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_CONSENSUS_PARAMS_ENABLE_TIMED_GEN_EMPTY_BLOCK=false
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_CONSENSUS_PARAMS_BLOCK_MAX_TX_NUM=500
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_CONSENSUS_PARAMS_HIGH_WATERMARK_CHECKPOINT_PERIOD=40
-#export AXIOM_LEDGER_GENESIS_EPOCH_INFO_CONSENSUS_PARAMS_PROPOSER_ELECTION_TYPE=0
+#export AXIOM_LEDGER_EPOCH_INFO_EPOCH_PERIOD=100
+#export AXIOM_LEDGER_EPOCH_INFO_CONSENSUS_PARAMS_CHECKPOINT_PERIOD=1
+#export AXIOM_LEDGER_EPOCH_INFO_CONSENSUS_PARAMS_ENABLE_TIMED_GEN_EMPTY_BLOCK=false
+#export AXIOM_LEDGER_EPOCH_INFO_CONSENSUS_PARAMS_BLOCK_MAX_TX_NUM=500
+#export AXIOM_LEDGER_EPOCH_INFO_CONSENSUS_PARAMS_HIGH_WATERMARK_CHECKPOINT_PERIOD=40
+#export AXIOM_LEDGER_EPOCH_INFO_CONSENSUS_PARAMS_PROPOSER_ELECTION_TYPE=0
 #export AXIOM_LEDGER_TX_CACHE_SET_SIZE=50
+#export AXIOM_LEDGER_BALANCE=200000000000000000000000000000
 
 
 # log
